+++ src/pycheval/const.py
@@ -16,6 +16,7 @@ URN_XRECHNUNG_PROFILE = "urn:cen.eu:en16931:2017#compliant#urn:xoev-de:kosit:sta
 NS_CII = "urn:un:unece:uncefact:data:standard:CrossIndustryInvoice:100"
 NS_RAM = "urn:un:unece:uncefact:data:standard:ReusableAggregateBusinessInformationEntity:100"  # noqa: E501
 NS_UDT = "urn:un:unece:uncefact:data:standard:UnqualifiedDataType:100"
+NS_QDT = "urn:un:unece:uncefact:data:standard:QualifiedDataType:100"
 
 ALLOWED_ATTACHMENT_MIME_TYPES = (
     "application/pdf",